pyproject.toml
[project]
name = "ice-node-repro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.4",
    "pyhealth==1.1.4",
    "torchdiffeq>=0.2.5",
]
